+++ packages/yourloops/components/patient-list/models/grid-row.model.ts
@@ -25,7 +25,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { type PatientListColumns } from '../enums/patient-list.enum'
+import { type PatientListColumns } from './enums/patient-list.enum'
 import { type Patient } from '../../../lib/patient/models/patient.model'
 import { type MonitoringAlerts } from '../../../lib/patient/models/monitoring-alerts.model'
 
@@ -34,9 +34,7 @@ export interface GridRowModel {
   [PatientListColumns.Flag]: Patient
   [PatientListColumns.Patient]: Patient
   [PatientListColumns.System]: string
-  [PatientListColumns.TimeOutOfRange]: MonitoringAlerts
-  [PatientListColumns.SevereHypoglycemia]: MonitoringAlerts
-  [PatientListColumns.DataNotTransferred]: MonitoringAlerts
+  [PatientListColumns.MonitoringAlerts]?: MonitoringAlerts
   [PatientListColumns.LastDataUpdate]: string
   [PatientListColumns.Messages]: boolean
   [PatientListColumns.Actions]: Patient